use explicit hint tokens for export marker
Since unary operator processing in external scanner conflicts with
export marker, the export marker was made into a non-terminal rule that
was also an external symbol, where the sole reason for the external
symbol was to allow the unary operator handler in the external parser
know that the marker `*` is meant to be a marker and not an unary
operator.

Recent tree-sitter changes has forbade this particular usage. As such,
switch to using a signal token (an optional token that is an external)
to let the external scanner know the context it is handling.

Ref: https://github.com/alaviss/tree-sitter-nim/issues/90#issuecomment-2211742258
Ref: https://github.com/tree-sitter/tree-sitter/pull/2577